requirements.txt
streamlit
pandas
openpyxl
python-dotenv
supabase
postgrest
